% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_provider.R
\name{configure_provider}
\alias{configure_provider}
\title{Configure Provider}
\usage{
configure_provider(name, api_key)
}
\arguments{
\item{name}{Name of the provider (one of "google_gemini" or "anthropic").}

\item{api_key}{The provider's API key to use for authentication. If `NULL`, the provider will be
deleted.}
}
\description{
Set API keys for running models locally.
}
\examples{
\dontrun{
configure_provider("google_gemini", "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
configure_provider(
  "anthropic",
  "sk-ant-api03-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
)
configure_provider("google_gemini", NULL)
}

}
