% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_print.R
\name{debug_print}
\alias{debug_print}
\title{Debug Print Function}
\usage{
debug_print(value)
}
\arguments{
\item{value}{The R object to be printed.}
}
\description{
Prints a value to the console with a timestamp if the 'DEBUG' environment variable is set.
This function is useful for conditional debugging output in R projects.
}
\keyword{internal}
