% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chat_instance.R
\name{get_chat_instance}
\alias{get_chat_instance}
\title{Get the ellmer Chat Instance}
\usage{
get_chat_instance(provider, model, api_key, mode, project_context)
}
\arguments{
\item{provider}{The model provider (E.g., "google_gemini", "anthropic").}

\item{model}{The ID of the model to use (E.g., "gemini-2.5-pro").}

\item{api_key}{The provider's API key to use for authentication.}

\item{mode}{The mode of operation, one of "agent" or "ask".}

\item{project_context}{The context of the session executing the addin, obtained with
`get_project_context()`.}
}
\description{
Get the ellmer Chat Instance
}
\keyword{internal}
