% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get a Configuration Value}
\usage{
get_config(config)
}
\arguments{
\item{config}{The name of the configuration value to obtain.}
}
\description{
Get a Configuration Value
}
\keyword{internal}
