% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_models.R
\name{get_ellmer_models}
\alias{get_ellmer_models}
\title{Get Ellmer Models}
\usage{
get_ellmer_models(provider, api_key)
}
\arguments{
\item{provider}{A character string indicating the model provider (e.g., "anthropic", "deepseek",
"google_gemini", "ollama", "openai").}

\item{api_key}{A character string containing the API key or credentials used to query the
provider.}
}
\description{
Get Ellmer Models
}
\keyword{internal}
