% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_models.R
\name{get_myownrobs_models}
\alias{get_myownrobs_models}
\title{Get MyOwnRobs Models}
\usage{
get_myownrobs_models(api_url)
}
\arguments{
\item{api_url}{The API URL to use for requests.}
}
\description{
Get MyOwnRobs Models
}
\keyword{internal}
