% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_context.R
\name{get_project_context}
\alias{get_project_context}
\title{Get Project Context Information}
\usage{
get_project_context()
}
\description{
Gathers various pieces of context information about the current R session and project. It
includes details about the working directory, RStudio project, source editor context, and
operating system.
}
\keyword{internal}
