% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{get_turn_role}
\alias{get_turn_role}
\title{Get the MyOwnRobs Role from a Turn}
\usage{
get_turn_role(turn)
}
\arguments{
\item{turn}{A Turn.}
}
\description{
Get the MyOwnRobs Role from a Turn
}
\keyword{internal}
