% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_models.R
\name{models_deepseek}
\alias{models_deepseek}
\title{Retrieve Available DeepSeek Models}
\usage{
models_deepseek(credentials = NULL)
}
\arguments{
\item{credentials}{A function that returns the API key as a character string.}
}
\description{
Fetches a list of currently available models from the DeepSeek API.
}
\keyword{internal}
