% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownhadley.R
\name{myownhadley_server}
\alias{myownhadley_server}
\title{MyOwnHadley Shiny Server}
\usage{
myownhadley_server(api_url)
}
\arguments{
\item{api_url}{The API URL to use for requests.}
}
\description{
MyOwnHadley Shiny Server
}
\keyword{internal}
