% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs_ui}
\alias{myownrobs_ui}
\title{MyOwnRobs Shiny UI}
\usage{
myownrobs_ui(available_models)
}
\arguments{
\item{available_models}{List of available models to use.}
}
\description{
MyOwnRobs Shiny UI
}
\keyword{internal}
