% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_models.R
\name{nice_names}
\alias{nice_names}
\title{Make Nicer Names}
\usage{
nice_names(names)
}
\arguments{
\item{names}{A character vector of names to be converted to a nicer, title-cased form.}
}
\description{
Make Nicer Names
}
\keyword{internal}
