% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_policy_acceptance.R
\name{policy_server}
\alias{policy_server}
\title{Policy Acceptance Shiny Server}
\usage{
policy_server(policy)
}
\arguments{
\item{policy}{A list with the policy.}
}
\description{
Policy Acceptance Shiny Server
}
\keyword{internal}
