% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_policy_acceptance.R
\name{policy_ui}
\alias{policy_ui}
\title{Policy Acceptance Shiny UI}
\usage{
policy_ui(policy)
}
\arguments{
\item{policy}{A list with the policy.}
}
\description{
Policy Acceptance Shiny UI
}
\keyword{internal}
