% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{save_turns}
\alias{save_turns}
\title{Set a Configuration Value}
\usage{
save_turns(value)
}
\arguments{
\item{value}{The value to assign to assign to the configuration.}

\item{config}{The name of the configuration value to set.}
}
\description{
Set a Configuration Value
}
\keyword{internal}
