% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_prompt.R
\name{send_prompt}
\alias{send_prompt}
\title{Send Prompt to the LLM}
\usage{
send_prompt(chat_id, prompt, role, mode, model, api_url)
}
\arguments{
\item{chat_id}{The ID of the chat session.}

\item{prompt}{The prompt to send.}

\item{role}{The role of the entity sending the prompt, one of "user" or "tool_runner".}

\item{mode}{The mode of operation, one of "agent" or "ask".}

\item{model}{The ID of the model to use.}

\item{api_url}{The API URL to use for requests.}
}
\description{
Send Prompt to the LLM
}
