% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_prompt.R
\name{send_prompt_async}
\alias{send_prompt_async}
\title{Asynchronously Send Prompt to the LLM}
\usage{
send_prompt_async(
  prompt,
  mode,
  model,
  project_context,
  api_key,
  available_models
)
}
\arguments{
\item{prompt}{The prompt to send.}

\item{mode}{The mode of operation, one of "agent" or "ask".}

\item{model}{The ID of the model to use (E.g., "gemini-2.5-pro").}

\item{project_context}{The context of the session executing the addin, obtained with
`get_project_context()`.}

\item{api_key}{The API key for MyOwnRobs, obtained with `get_api_key()`.}

\item{available_models}{List of available models to use.}
}
\description{
Asynchronously Send Prompt to the LLM
}
\keyword{internal}
