% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Set a Configuration Value}
\usage{
set_config(config, value)
}
\arguments{
\item{config}{The name of the configuration value to set.}

\item{value}{The value to assign to the configuration.}
}
\description{
Set a Configuration Value
}
\keyword{internal}
