% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_initial_project.R
\name{set_initial_project}
\alias{set_initial_project}
\title{Set the Initial State of the Project}
\usage{
set_initial_project(restore = FALSE)
}
\arguments{
\item{restore}{If TRUE, it will restore the initial state. If FALSE, it will save the initial
store in R's temp dir.}
}
\description{
Whether to save the initial state or to restore it.
}
\keyword{internal}
