% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_module}
\alias{settings_module}
\title{Settings Module}
\usage{
settings_module(id, r_trigger)
}
\arguments{
\item{id}{Module id.}

\item{r_trigger}{A reactive expression that triggers opening the modal.}
}
\description{
Settings Module
}
\keyword{internal}
