% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{turns_to_ui}
\alias{turns_to_ui}
\title{Convert a Turns Structure Into a User-Readable UI}
\usage{
turns_to_ui(turns)
}
\arguments{
\item{turns}{A list of Turns.}
}
\description{
Convert a Turns Structure Into a User-Readable UI
}
\keyword{internal}
