% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_commands.R
\name{validate_command_args}
\alias{validate_command_args}
\title{Validate Command Arguments}
\usage{
validate_command_args(command, args)
}
\arguments{
\item{command}{A list representing the command definition, expected to contain a 'parameters'
element.}

\item{args}{A list of arguments provided to the command.}
}
\description{
Checks if all required parameters for a given command are present in the provided arguments.
This function ensures that the AI agent has supplied all necessary arguments before executing a
command.
}
\keyword{internal}
