% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_credentials.R
\name{validate_credentials}
\alias{validate_credentials}
\title{Validate MyOwnRobs Credentials}
\usage{
validate_credentials(api_url, force = FALSE)
}
\arguments{
\item{api_url}{The API URL to use for requests.}

\item{force}{Force validation altough there's an existing key.}
}
\description{
Validate MyOwnRobs Credentials
}
\keyword{internal}
