% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_policy_acceptance.R
\name{validate_policy_acceptance}
\alias{validate_policy_acceptance}
\title{Validate MyOwnRobs Policy Acceptance}
\usage{
validate_policy_acceptance()
}
\description{
Checks if the user already accepted the usage policy. If the user didn't it prompts the policies.
}
\keyword{internal}
