% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turns.R
\name{content_to_ui}
\alias{content_to_ui}
\title{Convert a Turn Content into a User-Readable UI}
\usage{
content_to_ui(content, role)
}
\arguments{
\item{content}{A Content.}

\item{role}{The Turn's main role.}
}
\description{
Convert a Turn Content into a User-Readable UI
}
\keyword{internal}
