% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownhadley.R
\name{myownhadley}
\alias{myownhadley}
\title{Launch MyOwnHadley}
\usage{
myownhadley(
  api_url = paste0("https://myownhadley.com/api/v", packageVersion("myownhadley")$major)
)
}
\arguments{
\item{api_url}{The API URL to use for requests.}
}
\description{
Open a Shiny app addin with the chat interface.
}
