% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs}
\alias{myownrobs}
\title{Launch MyOwnRobs}
\usage{
myownrobs(
  api_url = paste0("https://myownhadley.com/api/v", packageVersion("myownrobs")$major)
)
}
\arguments{
\item{api_url}{The API URL to use for requests. This parameter is for advanced users who want to
specify an alternative backend URL and is rarely needed.}
}
\value{
No return value. Called for its side effects to launch the MyOwnRobs RStudio addin.
}
\description{
Open the RStudio addin with the chat interface.
}
\examples{
if (interactive()) {
  myownrobs()
  # Specify the API URL.
  myownrobs("https://myownhadley.com/api/v0")
}

}
