% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs}
\alias{myownrobs}
\title{Launch MyOwnRobs}
\usage{
myownrobs(
  api_url = paste0("https://myownhadley.com/api/v", packageVersion("myownrobs")$major)
)
}
\arguments{
\item{api_url}{The API URL to use for requests.}
}
\value{
No return value. Called for its side effects to launch the MyOwnRobs RStudio addin.
}
\description{
Open the RStudio addin with the chat interface.
}
\examples{
if (interactive()) {
  myownrobs()
}

}
