% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs}
\alias{myownrobs}
\title{Launch MyOwnRobs}
\usage{
myownrobs()
}
\value{
No return value. Called for its side effects to launch the MyOwnRobs RStudio addin.
}
\description{
Open the RStudio addin with the chat interface.
}
\examples{
if (interactive()) {
  # Configure your API providers first.
  configure_provider("google_gemini", Sys.getenv("GEMINI_API_KEY"))
  # Then launch MyOwnRobs.
  myownrobs()
}

}
