% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs_server}
\alias{myownrobs_server}
\title{MyOwnRobs Shiny Server}
\usage{
myownrobs_server(api_url)
}
\arguments{
\item{api_url}{The API URL to use for requests.}
}
\description{
MyOwnRobs Shiny Server
}
\keyword{internal}
