% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myownrobs.R
\name{myownrobs_server}
\alias{myownrobs_server}
\title{MyOwnRobs Shiny Server}
\usage{
myownrobs_server(available_models, project_context)
}
\arguments{
\item{available_models}{List of available models to use, obtained with `get_available_models()`.}

\item{project_context}{The context of the session executing the addin, obtained with
`get_project_context()`.}
}
\description{
MyOwnRobs Shiny Server
}
\keyword{internal}
