% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_credentials.R
\name{save_api_key}
\alias{save_api_key}
\title{Save MyOwnRobs Credentials Locally}
\usage{
save_api_key(api_key)
}
\arguments{
\item{api_key}{The MyOwnRobs API key to save locally.}
}
\description{
Save MyOwnRobs Credentials Locally
}
\keyword{internal}
